add templates everywhere and get them to compile
