+++ Layer.h
@@ -3,11 +3,11 @@
 
 #include "MatrixCommon.h"
 
+template <typename RGB>
 class SM_Layer {
     public:
         virtual void frameRefreshCallback();
-        virtual void getRefreshPixel(uint8_t x, uint8_t y, rgb24 &refreshPixel);
-        virtual void getRefreshPixel(uint8_t x, uint8_t y, rgb48 &refreshPixel);
+        virtual void getRefreshPixel(uint8_t x, uint8_t y, RGB &refreshPixel);
         void setRotation(rotationDegrees newrotation);
         void setRefreshRate(uint8_t newRefreshRate);
         void addLayer(SM_Layer * newlayer);
@@ -20,5 +20,7 @@ class SM_Layer {
         
     private:
 };
+
+#include "Layer_Impl.h"
 
 #endif
