+++ Layer_Foreground_Impl.h
@@ -1,44 +1,46 @@
-#include "Layer_Foreground.h"
 #include <string.h>
 
 const unsigned char foregroundDrawBuffer = 0;
 const unsigned char foregroundRefreshBuffer = 1;
 
-#define FOREGROUND_ROW_SIZE     (localWidth / 8)
-#define FOREGROUND_BUFFER_SIZE  (FOREGROUND_ROW_SIZE * localHeight)
+#define FOREGROUND_ROW_SIZE     (this->localWidth / 8)
+#define FOREGROUND_BUFFER_SIZE  (FOREGROUND_ROW_SIZE * this->localHeight)
 
-SMLayerForeground::SMLayerForeground(uint8_t * bitmap, uint8_t width, uint8_t height) {
+template <typename RGB>
+SMLayerForeground<RGB>::SMLayerForeground(uint8_t * bitmap, uint8_t width, uint8_t height) {
     // size of bitmap is 2 * FOREGROUND_BUFFER_SIZE
     foregroundBitmap = bitmap;
-    matrixWidth = width;
-    matrixHeight = height;
+    this->matrixWidth = width;
+    this->matrixHeight = height;
 }
 
-void SMLayerForeground::frameRefreshCallback(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::frameRefreshCallback(void) {
     handleForegroundDrawingCopy();
     updateForeground();
 }
 
 // returns true and copies color to xyPixel if pixel is opaque, returns false if not
-bool SMLayerForeground::getForegroundPixel(uint8_t hardwareX, uint8_t hardwareY, rgb24 &xyPixel) {
+template<typename RGB>
+bool SMLayerForeground<RGB>::getForegroundPixel(uint8_t hardwareX, uint8_t hardwareY, RGB &xyPixel) {
     uint8_t localScreenX, localScreenY;
 
     // convert hardware x/y to the pixel in the local screen
-    switch( rotation ) {
+    switch( this->rotation ) {
       case rotation0 :
         localScreenX = hardwareX;
         localScreenY = hardwareY;
         break;
       case rotation180 :
-        localScreenX = (matrixWidth - 1) - hardwareX;
-        localScreenY = (matrixHeight - 1) - hardwareY;
+        localScreenX = (this->matrixWidth - 1) - hardwareX;
+        localScreenY = (this->matrixHeight - 1) - hardwareY;
         break;
       case  rotation90 :
         localScreenX = hardwareY;
-        localScreenY = (matrixWidth - 1) - hardwareX;
+        localScreenY = (this->matrixWidth - 1) - hardwareX;
         break;
       case  rotation270 :
-        localScreenX = (matrixHeight - 1) - hardwareY;
+        localScreenX = (this->matrixHeight - 1) - hardwareY;
         localScreenY = hardwareX;
         break;
       default:
@@ -49,15 +51,16 @@ bool SMLayerForeground::getForegroundPixel(uint8_t hardwareX, uint8_t hardwareY,
     uint8_t bitmask = 0x80 >> (localScreenX % 8);
 
     if (foregroundBitmap[(foregroundRefreshBuffer * FOREGROUND_BUFFER_SIZE) + (localScreenY * FOREGROUND_ROW_SIZE) + (localScreenX/8)] & bitmask) {
-        copyRgb24(xyPixel, textcolor);
+        xyPixel = textcolor;
         return true;
     }
 
     return false;
 }
 
-void SMLayerForeground::getRefreshPixel(uint8_t hardwareX, uint8_t hardwareY, rgb48 &xyPixel) {
-    rgb24 tempPixel;
+template<typename RGB>
+void SMLayerForeground<RGB>::getRefreshPixel(uint8_t hardwareX, uint8_t hardwareY, RGB &xyPixel) {
+    RGB tempPixel;
 
     // do once per refresh
     bool bHasCC = ccmode != ccNone;
@@ -77,37 +80,19 @@ void SMLayerForeground::getRefreshPixel(uint8_t hardwareX, uint8_t hardwareY, rg
     }
 }
 
-void SMLayerForeground::getRefreshPixel(uint8_t hardwareX, uint8_t hardwareY, rgb24 &xyPixel) {
-    rgb24 tempPixel;
-
-    // do once per refresh
-    bool bHasCC = ccmode != ccNone;
-
-    if(getForegroundPixel(hardwareX, hardwareY, tempPixel)) {
-        if(bHasCC) {
-            // load foreground pixel with color correction
-            xyPixel.red = colorCorrection(tempPixel.red);
-            xyPixel.green = colorCorrection(tempPixel.green);
-            xyPixel.blue = colorCorrection(tempPixel.blue);
-        } else {
-            // load foreground pixel without color correction
-            xyPixel.red = tempPixel.red;
-            xyPixel.green = tempPixel.green;
-            xyPixel.blue = tempPixel.blue;
-        }
-    }
+template<typename RGB>
+void SMLayerForeground<RGB>::setScrollColor(const RGB & newColor) {
+    textcolor = newColor;
 }
 
-void SMLayerForeground::setScrollColor(const rgb24 & newColor) {
-    copyRgb24(textcolor, newColor);
-}
-
-void SMLayerForeground::setColorCorrection(colorCorrectionModes mode) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setColorCorrection(colorCorrectionModes mode) {
     ccmode = mode;
 }
 
 // stops the scrolling text on the next refresh
-void SMLayerForeground::stopScrollText(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::stopScrollText(void) {
     // setup conditions for ending scrolling:
     // scrollcounter is next to zero
     scrollcounter = 1;
@@ -115,11 +100,13 @@ void SMLayerForeground::stopScrollText(void) {
     scrollPosition = scrollMin;
 }
 
-void SMLayerForeground::clearForeground(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::clearForeground(void) {
     memset(&foregroundBitmap[foregroundDrawBuffer*FOREGROUND_BUFFER_SIZE], 0x00, FOREGROUND_BUFFER_SIZE);
 }
 
-void SMLayerForeground::displayForegroundDrawing(bool waitUntilComplete) {
+template <typename RGB>
+void SMLayerForeground<RGB>::displayForegroundDrawing(bool waitUntilComplete) {
     while (foregroundCopyPending);
 
     foregroundCopyPending = true;
@@ -127,7 +114,8 @@ void SMLayerForeground::displayForegroundDrawing(bool waitUntilComplete) {
     while (waitUntilComplete && foregroundCopyPending);
 }
 
-void SMLayerForeground::handleForegroundDrawingCopy(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::handleForegroundDrawingCopy(void) {
     if (!foregroundCopyPending)
         return;
 
@@ -136,10 +124,11 @@ void SMLayerForeground::handleForegroundDrawingCopy(void) {
     foregroundCopyPending = false;
 }
 
-void SMLayerForeground::drawForegroundPixel(int16_t x, int16_t y, bool opaque) {
+template <typename RGB>
+void SMLayerForeground<RGB>::drawForegroundPixel(int16_t x, int16_t y, bool opaque) {
     uint8_t tempBitmask;
 
-    if(x < 0 || x >= localWidth || y < 0 || y >= localWidth)
+    if(x < 0 || x >= this->localWidth || y < 0 || y >= this->localWidth)
         return;
 
     if(opaque) {
@@ -151,24 +140,26 @@ void SMLayerForeground::drawForegroundPixel(int16_t x, int16_t y, bool opaque) {
     }
 }
 
-void SMLayerForeground::setForegroundFont(fontChoices newFont) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setForegroundFont(fontChoices newFont) {
     foregroundfont = (bitmap_font *)fontLookup(newFont);
     majorScrollFontChange = true;
 }
 
-void SMLayerForeground::drawForegroundChar(int16_t x, int16_t y, char character, bool opaque) {
+template <typename RGB>
+void SMLayerForeground<RGB>::drawForegroundChar(int16_t x, int16_t y, char character, bool opaque) {
     uint8_t tempBitmask;
     int k;
 
     // only draw if character is on the screen
-    if (x + scrollFont->Width < 0 || x >= localWidth) {
+    if (x + scrollFont->Width < 0 || x >= this->localWidth) {
         return;
     }
 
     for (k = y; k < y+foregroundfont->Height; k++) {
         // ignore rows that are not on the screen
         if(k < 0) continue;
-        if (k >= localHeight) return;
+        if (k >= this->localHeight) return;
 
         tempBitmask = getBitmapFontRowAtXY(character, k - y, foregroundfont);
         if (x < 0) {
@@ -176,13 +167,14 @@ void SMLayerForeground::drawForegroundChar(int16_t x, int16_t y, char character,
         } else {
             foregroundBitmap[foregroundDrawBuffer*FOREGROUND_BUFFER_SIZE + (k * FOREGROUND_ROW_SIZE) + (x/8)] |= tempBitmask >> (x%8);
             // do two writes if the shifted 8-bit wide bitmask is still on the screen
-            if(x + 8 < localWidth && x % 8)
+            if(x + 8 < this->localWidth && x % 8)
                 foregroundBitmap[foregroundDrawBuffer*FOREGROUND_BUFFER_SIZE + (k * FOREGROUND_ROW_SIZE) + (x/8) + 1] |= tempBitmask << (8-(x%8));
         }
     }
 }
 
-void SMLayerForeground::drawForegroundString(int16_t x, int16_t y, const char text [], bool opaque) {
+template <typename RGB>
+void SMLayerForeground<RGB>::drawForegroundString(int16_t x, int16_t y, const char text [], bool opaque) {
     // limit text to 10 chars, why?
     for (int i = 0; i < 10; i++) {
         char character = text[i];
@@ -193,7 +185,8 @@ void SMLayerForeground::drawForegroundString(int16_t x, int16_t y, const char te
     }
 }
 
-void SMLayerForeground::drawForegroundMonoBitmap(int16_t x, int16_t y, uint8_t width, uint8_t height, uint8_t *bitmap, bool opaque) {
+template <typename RGB>
+void SMLayerForeground<RGB>::drawForegroundMonoBitmap(int16_t x, int16_t y, uint8_t width, uint8_t height, uint8_t *bitmap, bool opaque) {
     int xcnt, ycnt;
 
     for (ycnt = 0; ycnt < height; ycnt++) {
@@ -208,18 +201,20 @@ void SMLayerForeground::drawForegroundMonoBitmap(int16_t x, int16_t y, uint8_t w
 // returns 0 if stopped
 // returns positive number indicating number of loops left if running
 // returns -1 if continuously scrolling
-int SMLayerForeground::getScrollStatus(void) const {
+template <typename RGB>
+int SMLayerForeground<RGB>::getScrollStatus(void) const {
     return scrollcounter;
 }
 
-void SMLayerForeground::setScrollMinMax(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollMinMax(void) {
    switch (scrollmode) {
     case wrapForward:
     case bounceForward:
     case bounceReverse:
     case wrapForwardFromLeft:
         scrollMin = -textWidth;
-        scrollMax = localWidth;
+        scrollMax = this->localWidth;
 
         scrollPosition = scrollMax;
 
@@ -240,7 +235,8 @@ void SMLayerForeground::setScrollMinMax(void) {
 
 }
 
-void SMLayerForeground::scrollText(const char inputtext[], int numScrolls) {
+template <typename RGB>
+void SMLayerForeground<RGB>::scrollText(const char inputtext[], int numScrolls) {
     int length = strlen((const char *)inputtext);
     if (length > textLayerMaxStringLength)
         length = textLayerMaxStringLength;
@@ -255,7 +251,8 @@ void SMLayerForeground::scrollText(const char inputtext[], int numScrolls) {
 
 //Updates the text that is currently scrolling to the new value
 //Useful for a clock display where the time changes.
-void SMLayerForeground::updateScrollText(const char inputtext[]){
+template <typename RGB>
+void SMLayerForeground<RGB>::updateScrollText(const char inputtext[]){
     int length = strlen((const char *)inputtext);
     if (length > textLayerMaxStringLength)
         length = textLayerMaxStringLength;
@@ -268,7 +265,8 @@ void SMLayerForeground::updateScrollText(const char inputtext[]){
 
 // called once per frame to update foreground (virtual) bitmap
 // function needs major efficiency improvments
-void SMLayerForeground::updateForeground(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::updateForeground(void) {
     bool resetScrolls = false;
 
     // return if not ready to update
@@ -324,38 +322,44 @@ void SMLayerForeground::updateForeground(void) {
 }
 
 // TODO: recompute stuff after changing mode, font, etc
-void SMLayerForeground::setScrollMode(ScrollMode mode) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollMode(ScrollMode mode) {
     scrollmode = mode;
 }
 
 // TODO:need to get refresh rate from main class
 
 
-void SMLayerForeground::setScrollSpeed(unsigned char pixels_per_second) {
-    framesperscroll = (refreshRate * 1.0) / pixels_per_second;
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollSpeed(unsigned char pixels_per_second) {
+    framesperscroll = (this->refreshRate * 1.0) / pixels_per_second;
 }
 
-void SMLayerForeground::setScrollFont(fontChoices newFont) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollFont(fontChoices newFont) {
     scrollFont = fontLookup(newFont);
 }
 
-void SMLayerForeground::setScrollOffsetFromTop(int offset) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollOffsetFromTop(int offset) {
     fontTopOffset = offset;
     majorScrollFontChange = true;
 }
 
-void SMLayerForeground::setScrollStartOffsetFromLeft(int offset) {
+template <typename RGB>
+void SMLayerForeground<RGB>::setScrollStartOffsetFromLeft(int offset) {
     fontLeftOffset = offset;
 }
 
 // if font size or position changed since the last call, redraw the whole frame
-void SMLayerForeground::redrawForeground(void) {
+template <typename RGB>
+void SMLayerForeground<RGB>::redrawForeground(void) {
     int j, k;
     int charPosition, textPosition;
     uint8_t charY0, charY1;
 
 
-    for (j = 0; j < localHeight; j++) {
+    for (j = 0; j < this->localHeight; j++) {
 
         // skip rows without text
         if (j < fontTopOffset || j >= fontTopOffset + scrollFont->Height)
@@ -375,8 +379,8 @@ void SMLayerForeground::redrawForeground(void) {
         // find rows within character bitmap that will be drawn (0-font->height unless text is partially off screen)
         charY0 = j - fontTopOffset;
 
-        if (localHeight < fontTopOffset + scrollFont->Height) {
-            charY1 = localHeight - fontTopOffset;
+        if (this->localHeight < fontTopOffset + scrollFont->Height) {
+            charY1 = this->localHeight - fontTopOffset;
         } else {
             charY1 = scrollFont->Height;
         }
@@ -396,7 +400,7 @@ void SMLayerForeground::redrawForeground(void) {
         for (k = 0; k < charY1 - charY0; k++)
             memset(&foregroundBitmap[foregroundRefreshBuffer*FOREGROUND_BUFFER_SIZE + ((j + k) * FOREGROUND_ROW_SIZE)], 0x00, FOREGROUND_ROW_SIZE);
 
-        while (textPosition < textlen && charPosition < localWidth) {
+        while (textPosition < textlen && charPosition < this->localWidth) {
             uint8_t tempBitmask;
             // draw character from top to bottom
             for (k = charY0; k < charY1; k++) {
@@ -407,7 +411,7 @@ void SMLayerForeground::redrawForeground(void) {
                 } else {
                     foregroundBitmap[foregroundRefreshBuffer*FOREGROUND_BUFFER_SIZE + ((j + k - charY0) * FOREGROUND_ROW_SIZE) + (charPosition/8)] |= tempBitmask >> (charPosition%8);
                     // do two writes if the shifted 8-bit wide bitmask is still on the screen
-                    if(charPosition + 8 < localWidth && charPosition % 8)
+                    if(charPosition + 8 < this->localWidth && charPosition % 8)
                         foregroundBitmap[foregroundRefreshBuffer*FOREGROUND_BUFFER_SIZE + ((j + k - charY0) * FOREGROUND_ROW_SIZE) + (charPosition/8) + 1] |= tempBitmask << (8-(charPosition%8));
                 }
             }
@@ -421,7 +425,8 @@ void SMLayerForeground::redrawForeground(void) {
     }
 }
 
-bool SMLayerForeground::getBitmapPixelAtXY(uint8_t x, uint8_t y, uint8_t width, uint8_t height, const uint8_t *bitmap) {
+template <typename RGB>
+bool SMLayerForeground<RGB>::getBitmapPixelAtXY(uint8_t x, uint8_t y, uint8_t width, uint8_t height, const uint8_t *bitmap) {
     int cell = (y * ((width / 8) + 1)) + (x / 8);
 
     uint8_t mask = 0x80 >> (x % 8);
