+++ SmartMatrix.h
@@ -39,7 +39,7 @@
 //#if COLOR_DEPTH_RGB > 24
 typedef rgb48 refreshPixel;
 //#else
-//typedef rgb24 refreshPixel;
+//typedef RGB refreshPixel;
 //#endif
 
 // definition telling FastLED that setBackBuffer() and getRealBackBuffer() are available
@@ -63,6 +63,7 @@ typedef struct matrixUpdateBlock {
     addresspair addressValues;
 } matrixUpdateBlock;
 
+template <typename RGB>
 class SmartMatrix {
 public:
     SmartMatrix(uint8_t width, uint8_t height, uint8_t depth, uint8_t bufferrows, uint32_t * dataBuffer, uint8_t * blockBuffer);
@@ -74,41 +75,44 @@ public:
     void swapBuffersWithInterpolation_frames(int framesToInterpolate, bool copy = false);
     void swapBuffersWithInterpolation_ms(int interpolationSpan_ms, bool copy = false);
 #endif
-    void drawPixel(int16_t x, int16_t y, const rgb24& color);
-    void drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const rgb24& color);
-    void drawFastVLine(int16_t x, int16_t y0, int16_t y1, const rgb24& color);
-    void drawFastHLine(int16_t x0, int16_t x1, int16_t y, const rgb24& color);
-    void drawCircle(int16_t x0, int16_t y0, uint16_t radius, const rgb24& color);
-    void fillCircle(int16_t x0, int16_t y0, uint16_t radius, const rgb24& outlineColor, const rgb24& fillColor);
-    void fillCircle(int16_t x0, int16_t y0, uint16_t radius, const rgb24& color);
-    void drawEllipse(int16_t x0, int16_t y0, uint16_t radiusX, uint16_t radiusY, const rgb24& color);
-    void drawTriangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, int16_t x3, int16_t y3, const rgb24& color);
-    void fillTriangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, int16_t x3, int16_t y3, const rgb24& fillColor);
+    void drawPixel(int16_t x, int16_t y, const RGB& color);
+    void drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const RGB& color);
+    void drawFastVLine(int16_t x, int16_t y0, int16_t y1, const RGB& color);
+    void drawFastHLine(int16_t x0, int16_t x1, int16_t y, const RGB& color);
+    void drawCircle(int16_t x0, int16_t y0, uint16_t radius, const RGB& color);
+    void fillCircle(int16_t x0, int16_t y0, uint16_t radius, const RGB& outlineColor, const RGB& fillColor);
+    void fillCircle(int16_t x0, int16_t y0, uint16_t radius, const RGB& color);
+    void drawEllipse(int16_t x0, int16_t y0, uint16_t radiusX, uint16_t radiusY, const RGB& color);
+    void drawTriangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, int16_t x3, int16_t y3, const RGB& color);
+    void fillTriangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, int16_t x3, int16_t y3, const RGB& fillColor);
     void fillTriangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, int16_t x3, int16_t y3,
-    const rgb24& outlineColor, const rgb24& fillColor);
-    void drawRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const rgb24& color);
-    void fillRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const rgb24& color);
-    void fillRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const rgb24& outlineColor, const rgb24& fillColor);
-    void drawRoundRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t radius, const rgb24& outlineColor);
-    void fillRoundRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t radius, const rgb24& fillColor);
+        const RGB& outlineColor, const RGB& fillColor);
+    void drawRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const RGB& color);
+    void fillRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const RGB& color);
+    void fillRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const RGB& outlineColor, const RGB& fillColor);
+    void drawRoundRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t radius, const RGB& outlineColor);
+    void fillRoundRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t radius, const RGB& fillColor);
     void fillRoundRectangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t radius,
-    const rgb24& outlineColor, const rgb24& fillColor);
-    void fillScreen(const rgb24& color);
-    void drawChar(int16_t x, int16_t y, const rgb24& charColor, char character);
-    void drawString(int16_t x, int16_t y, const rgb24& charColor, const char text[]);
-    void drawString(int16_t x, int16_t y, const rgb24& charColor, const rgb24& backColor, const char text[]);
-    void drawMonoBitmap(int16_t x, int16_t y, uint8_t width, uint8_t height, const rgb24& bitmapColor, const uint8_t *bitmap);
-    const rgb24 readPixel(int16_t x, int16_t y);
-    rgb24 *backBuffer(void);
-    void setBackBuffer(rgb24 *newBuffer);
-    rgb24 *getRealBackBuffer(void);
+        const RGB& outlineColor, const RGB& fillColor);
+    void fillScreen(const RGB& color);
+    void drawChar(int16_t x, int16_t y, const RGB& charColor, char character);
+    void drawString(int16_t x, int16_t y, const RGB& charColor, const char text[]);
+    void drawString(int16_t x, int16_t y, const RGB& charColor, const RGB& backColor, const char text[]);
+    void drawMonoBitmap(int16_t x, int16_t y, uint8_t width, uint8_t height, const RGB& bitmapColor, const uint8_t *bitmap);
+
+    RGB readPixel(int16_t x, int16_t y);
+    RGB *backBuffer(void);
+
+    void setBackBuffer(RGB *newBuffer);
+
+    RGB *getRealBackBuffer();
 
     // scroll text
     void scrollText(const char inputtext[], int numScrolls);
     void setScrollMode(ScrollMode mode);
     void setScrollSpeed(unsigned char pixels_per_second);
     void setScrollFont(fontChoices newFont);
-    void setScrollColor(const rgb24 & newColor);
+    void setScrollColor(const RGB & newColor);
 #define setScrollOffsetFromEdge setScrollOffsetFromTop // backwards compatibility
     void setScrollOffsetFromTop(int offset);
     void setScrollStartOffsetFromLeft(int offset);
@@ -135,17 +139,19 @@ public:
     void setFont(fontChoices newFont);
     void setRefreshRate(uint8_t newRefreshRate);
 
-    SMLayerForeground * foregroundLayer;
-    SMLayerBackground * backgroundLayer;
+    SMLayerForeground<RGB> * foregroundLayer;
+    SMLayerBackground<RGB> * backgroundLayer;
 
     void useDefaultLayers(void);
-    void addLayer(SM_Layer * newlayer);
+    void addLayer(SM_Layer<RGB> * newlayer);
 
-    SM_Layer * baseLayer;
+    SM_Layer<RGB> * baseLayer;
 
 private:
     // enable ISR access to private member variables
+    template <typename RGB1>
     friend void rowCalculationISR(void);
+    template <typename RGB1>
     friend void rowShiftCompleteISR(void);
 
     // functions called by ISR
@@ -176,20 +182,22 @@ private:
     static matrixUpdateBlock * matrixUpdateBlocks;
     static addresspair * addressLUT;
     static timerpair * timerLUT;
+
+    static SmartMatrix<RGB>* globalinstance;
 };
 
 // single matrixUpdateBlocks buffer is divided up to hold matrixUpdateBlocks, addressLUT, timerLUT to simplify user sketch code and reduce constructor parameters
 #define SMARTMATRIX_ALLOCATE_BUFFERS(width, height, depth, rows) \
-static DMAMEM uint32_t matrixUpdateData[rows * width * (depth/3 / sizeof(uint32_t)) * 2]; \
-static DMAMEM uint8_t matrixUpdateBlocks[(sizeof(matrixUpdateBlock) * rows * depth/3) + (sizeof(addresspair) * height/2) + (sizeof(timerpair) * depth/3)]; \
-SmartMatrix matrix(width, height, depth, rows, matrixUpdateData, matrixUpdateBlocks)
+    static DMAMEM uint32_t matrixUpdateData[rows * width * (depth/3 / sizeof(uint32_t)) * 2]; \
+    static DMAMEM uint8_t matrixUpdateBlocks[(sizeof(matrixUpdateBlock) * rows * depth/3) + (sizeof(addresspair) * height/2) + (sizeof(timerpair) * depth/3)]; \
+    SmartMatrix<rgb##depth> matrix(width, height, depth, rows, matrixUpdateData, matrixUpdateBlocks)
 
-#define SMARTMATRIX_SETUP_DEFAULT_LAYERS(width, height) \
-    static rgb24 backgroundBitmap[2*width*height];                              \
-    static SMLayerBackground backgroundLayer(backgroundBitmap, width, height);  \
+#define SMARTMATRIX_SETUP_DEFAULT_LAYERS(width, height, depth) \
+    static rgb##depth backgroundBitmap[2*width*height];                              \
+    static SMLayerBackground<rgb##depth> backgroundLayer(backgroundBitmap, width, height);  \
     matrix.addLayer(&backgroundLayer);                                          \
     static uint8_t foregroundBitmap[2 * height * (width / 8)];                  \
-    static SMLayerForeground foregroundLayer(foregroundBitmap, width, height);  \
+    static SMLayerForeground<rgb##depth> foregroundLayer(foregroundBitmap, width, height);  \
     matrix.addLayer(&foregroundLayer);                                          \
     matrix.useDefaultLayers();                                                  \
     matrix.begin()
@@ -197,5 +205,10 @@ SmartMatrix matrix(width, height, depth, rows, matrixUpdateData, matrixUpdateBlo
 #define SMARTMATRIX_ALLOCATE_FOREGROUND_LAYER(layername, width, height) \
     static uint8_t foregroundBitmap[2 * width * (height / 8)];    \
     static SMLayerForeground layername(foregroundBitmap, width, height)
+
+#include "MatrixGraphics_Impl.h"
+#include "MatrixConfiguration_Impl.h"
+#include "MatrixForeground_Impl.h"
+#include "SmartMatrix_Impl.h"
 
 #endif
